Renamed files and image upload now updates user
